requirements.txt
# AI Trading Agent Dependencies

# Core AI and LangChain (compatible versions)
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
openai>=1.68.0

# Official Groww Trading API SDK
growwapi==0.0.7

# Data Analysis and Machine Learning (compatible with Python 3.13)
numpy>=1.24.0
scikit-learn>=1.3.0
yfinance>=0.2.0
ta>=0.10.0

# Configuration and Environment
python-dotenv==1.0.0
pydantic-settings>=2.0.0

# Scheduling and Task Management
schedule==1.2.0

# Web Scraping and Analysis
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Logging and Utilities
colorlog==6.8.0 